requirements.txt
gradio==4.8.0
langchain==0.0.349
openai==1.3.8
SocksiPy_branch==1.01
